+++ test/RewardsHandler.t.sol
@@ -924,16 +924,16 @@ contract TestableRewardsHandler is RewardsHandler, Staker, Depositor {
   // -------- Overridden abstract function placeholders --------
   function _updateUnstakesAfterTrigger(
     uint16, /* reservePoolId_ */
-    uint128, /* oldStakeAmount_ */
-    uint128 /* slashAmount_ */
+    uint256, /* oldStakeAmount_ */
+    uint256 /* slashAmount_ */
   ) internal view override {
     __readStub__();
   }
 
   function _updateWithdrawalsAfterTrigger(
     uint16, /* reservePoolId_ */
-    uint128, /* oldStakeAmount_ */
-    uint128 /* slashAmount_ */
+    uint256, /* oldStakeAmount_ */
+    uint256 /* slashAmount_ */
   ) internal view override {
     __readStub__();
   }
