+++ test/Staker.t.sol
@@ -347,16 +347,16 @@ contract TestableStaker is Staker, Depositor, RewardsHandler {
   // -------- Overridden abstract function placeholders --------
   function _updateWithdrawalsAfterTrigger(
     uint16, /* reservePoolId_ */
-    uint128, /* oldAmount_ */
-    uint128 /* slashAmount_ */
+    uint256, /* oldAmount_ */
+    uint256 /* slashAmount_ */
   ) internal view override {
     __readStub__();
   }
 
   function _updateUnstakesAfterTrigger(
     uint16, /* reservePoolId_ */
-    uint128, /* oldStakeAmount_ */
-    uint128 /* slashAmount_ */
+    uint256, /* oldStakeAmount_ */
+    uint256 /* slashAmount_ */
   ) internal view override {
     __readStub__();
   }
